requirements.txt
flask
vaderSentiment
cryptography
google-generativeai
